libsel4: Fix RISC-V type definitions

+++ libsel4/arch_include/riscv/sel4/arch/simple_types.h
@@ -21,11 +21,11 @@
 typedef signed char seL4_Int8;
 typedef signed short seL4_Int16;
 typedef signed int seL4_Int32;
-typedef signed long long seL4_Int64;
+typedef signed long seL4_Int64;
 
 typedef unsigned char seL4_Uint8;
 typedef unsigned short seL4_Uint16;
 typedef unsigned int seL4_Uint32;
-typedef unsigned long long seL4_Uint64;
+typedef unsigned long seL4_Uint64;
 
 #endif
